refactor 클래스 패키지별 분류
