+++ src/main/java/RacingGame/domain/Name.java
@@ -1,4 +1,4 @@
-package RacingGame;
+package RacingGame.domain;
 
 public class Name {
     private String name;
