+++ src/main/java/RacingGame/domain/Position.java
@@ -1,4 +1,4 @@
-package RacingGame;
+package RacingGame.domain;
 
 public class Position {
     private int position;
