+++ src/main/java/RacingGame/util/RandomNumber.java
@@ -1,4 +1,4 @@
-package RacingGame;
+package RacingGame.util;
 
 import java.util.Random;
