+++ src/main/java/RacingGame/view/InputView.java
@@ -1,4 +1,4 @@
-package RacingGame;
+package RacingGame.view;
 
 import java.util.Scanner;
