+++ src/test/java/RacingGameTest/RandomNumberTest.java
@@ -1,10 +1,8 @@
 package RacingGameTest;
 
-import RacingGame.RandomNumber;
+import RacingGame.util.RandomNumber;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-
-import java.util.Random;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
